renaming to happychat, moving ui state of happychat
