+++ client/state/action-types.js
@@ -88,7 +88,6 @@ export const HAPPYCHAT_SET_MESSAGE = 'HAPPYCHAT_SET_MESSAGE';
 export const HAPPYCHAT_RECEIVE_EVENT = 'HAPPYCHAT_RECEIVE_EVENT';
 export const HAPPYCHAT_SET_AUTOSCROLL = 'HAPPYCHAT_SET_AUTOSCROLL';
 export const HAPPYCHAT_OPEN = 'HAPPYCHAT_OPEN';
-export const HAPPYCHAT_OPEN_URL = 'HAPPYCHAT_OPEN_URL';
 export const IMAGE_EDITOR_CROP = 'IMAGE_EDITOR_CROP';
 export const IMAGE_EDITOR_FLIP = 'IMAGE_EDITOR_FLIP';
 export const IMAGE_EDITOR_ROTATE_COUNTERCLOCKWISE = 'IMAGE_EDITOR_ROTATE_COUNTERCLOCKWISE';
