+++ client/state/ui/happychat/selectors.js
@@ -1,0 +1,1 @@
+export const isHappychatOpen = ( state ) => state.ui.happychat.open && state.ui.section.name !== 'happychat'
